dev_requirements.txt
Flask
flask-cors
psutil
openai-whisper
numpy
SpeechRecognition
flask-socketio
torch

edge-tts
pygame
thefuzz
pyautogui
pygetwindow
screen_brightness_control